requirements.txt
anyio==3.6.2
certifi==2022.12.7
flet==0.5.0
flet-core==0.5.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
oauthlib==3.2.2
packaging==23.0
repath==0.9.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
speedtest-cli==2.1.3
SQLAlchemy==2.0.9
typing_extensions==4.5.0
watchdog==2.3.1
websocket-client==1.5.1
websockets==10.4
